pyproject.toml
[tool.poetry]
name = "srs-paper"
version = "0.1.0"
license = "MIT"
description = ""
authors = ["Todd Dembrey <todd.dembrey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.19.2"
pydicom = "^2.0.0"
Shapely = "^1.7.1"
pudb = "^2019.2"
matplotlib = "^3.3.2"
click = "^7.1.2"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
